Fix instruction pointer arithmetic
The instruction pointer was incorrectly being limited from being decremented to
the first instruction due to invalid limits checks. There were no tests
explicitly covering the behaviour of the instruction pointer manipulation as
this was an internal method and is not exposed in brianfuck outside of jump
instructions.

This branch corrects the bug, adds explicit tests for this behaviour to prevent
future regressions, and introduces some integration tests to ensure the language
is operating correcty.